workflow: Add workflow to attach release assets
When a release is created this new workflow calls the other build
and test workflows and uploads the tested artifacts into github
release.
The existing workflows were modified to make them callable from other
workflows.
Also added zephyr.signed.bin as an artifact so that it can be released to
github release.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>